+++ libs/core/src/lib/list/directives/list-icon.directive.ts
@@ -1,0 +1,48 @@
+import { Directive, ElementRef, Input, OnChanges, OnInit } from '@angular/core';
+import { applyCssClass } from '../../utils/decorators/apply-css-class.decorator';
+
+@Directive({
+  selector: '[fdListIcon], [fd-list-icon]'
+})
+export class ListIconDirective implements OnChanges, OnInit {
+
+    /** The icon name to display. See the icon page for the list of icons
+     * here: https://sap.github.io/fundamental-ngx/icon
+     * */
+    @Input()
+    glyph: string;
+
+    /** Apply user custom styles */
+    @Input()
+    class: string;
+
+    constructor(private _elementRef: ElementRef) {}
+
+    /** @hidden */
+    ngOnChanges(): void {
+        this.buildComponentCssClass();
+    }
+
+    /** @hidden */
+    ngOnInit(): void {
+        this.buildComponentCssClass();
+    }
+
+    @applyCssClass
+    /** CssClassBuilder interface implementation
+     * function must return single string
+     * function is responsible for order which css classes are applied
+     */
+    buildComponentCssClass(): string[] {
+        return [
+            'fd-list__icon',
+            this.glyph ? 'sap-icon--' + this.glyph : '',
+            this.class
+        ]
+    }
+
+    /** @hidden */
+    elementRef(): ElementRef<any> {
+        return this._elementRef;
+    }
+}